split logic for portion data accessor
